Move the `Metric` definition to its own file
