+++ glean/src/core/metrics/types/counter.ts
@@ -2,7 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Metric, MetricType, CommonMetricData } from "../index.js";
+import { MetricType, CommonMetricData } from "../index.js";
+import { Metric } from "../metric.js";
 import { isNumber, isUndefined, JSONValue } from "../../utils.js";
 import Glean from "../../glean.js";
