+++ glean/src/core/metrics/utils.ts
@@ -2,7 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Metric, LabeledMetric } from "./index.js";
+import { LabeledMetric } from "./index.js";
+import { Metric } from "./metric.js";
 import { JSONValue } from "../utils.js";
 
 import { BooleanMetric } from "./types/boolean.js";
